requirements.txt
python==3.8
tensorflow-gpu==2.6.0
cuDNN==8.1
CUDA==11.2
scenedetect==0.5.6.1

